Allow changing dimensions in Screen/Component
Previously, all values were absolute when given to Screen/Component
(x, y, w, h). This would break things if Screen/the app window ever
moved or was resized. Now the Screen/Component classes are able to
handle absolute values (always draw the x position as 20 to the left of
parent x), percent values (always draw the x position at 20% of the
width of parent). This broke the Yu-Gi-Oh Randomizer, so that will need
to be updated.