All float constructors added
Previously in the randomizer, the constructors for all Components was
all floats, the update I made to the library today made it so they
should all be Strings. Updated the library again so that there is a
separate constructor for all floats, which will convert them to Strings
with the "a" modifier, for absolute value.