Updated Component to hold instance of its own Screen
Components used to receive an instance of its own Screen every
draw/mouseOver function, which lead to slower performance in the YuGiOh
Randomizer program. Updated so that when a Component is added to a
Screen, the Screen gives it a copy of itself before adding the Component to its
ArrayList.